Refactor netty handlers.
1. Rename OvsdbClientHanlder to OvsdbConnectionHandler. It is now only
responsible for notify connections. Also merge HeartbeatHandler to this
handler.
2. Add JsonRpcHandler that handles the inbound JSON-RPC messages.
3. Add OvsdbClientImpl that can be constructed from a Netty channel.
